Refactored package structure of optical_tracer
Added /core with __init__ py.
Whole code from optical_tracer.py moved to various modules in /core.
Implemented facade in __init__.py using __all__.
Fixed imports in coupled modules in pytest, tkinter and django.
