+++ core/__init__.py
@@ -1,0 +1,11 @@
+from ._config import *
+from ._exceptions import *
+from ._optical_component import *
+from ._optical_system import *
+from ._rays import *
+
+__all__ = (_config.__all__ +
+           _exceptions.__all__ +
+           _optical_component.__all__ +
+           _optical_system.__all__ +
+           _rays.__all__)
